+++ Assets/스크립트/BT/가가가가가/CanMoveToPos.cs
@@ -7,30 +7,30 @@ using BehaviorDesigner.Runtime.Tasks;
 
 public class CanMoveToPos : Conditional
 {
+	InGameSceneIfRaidEnd inGameSceneIfRaidEnd;
 	InGameSceneCharSpineAniCon inGameSceneCharSpineAniCon;
 	CharState charState;
-	InGameSceneUiDataManager inGameSceneUiDataManager;
 
+	
 
 
 	public override void OnStart()
 	{
-		inGameSceneUiDataManager = GameObject.Find("Manager").gameObject.GetComponent<InGameSceneUiDataManager>();
 		inGameSceneCharSpineAniCon = this.gameObject.transform.GetComponent<InGameSceneCharSpineAniCon>();
+		inGameSceneIfRaidEnd = GameObject.Find("Manager").GetComponent<InGameSceneIfRaidEnd>();
 		charState = this.gameObject.transform.GetComponent<CharState>();
 	}
 
 	public override TaskStatus OnUpdate()
 	{
-		if(inGameSceneUiDataManager.waitForRaid == true &&
-				inGameSceneUiDataManager.isBattleStart == true)
+		if(inGameSceneIfRaidEnd.waitForRaid == true)
         {
 			charState.nowState = CharState.NowState.isWalkToOrginPos;
 			inGameSceneCharSpineAniCon.run();
 		}
 		else
         {
-		    ////charState.nowState = CharState.NowState.isIdle;
+			charState.nowState = CharState.NowState.isIdle;
 			inGameSceneCharSpineAniCon.idle();
 			return TaskStatus.Failure;
 		}
