+++ Assets/스크립트/BT/가가가가가/CanNormalAttack.cs
@@ -21,34 +21,41 @@ using BehaviorDesigner.Runtime.Tasks;
 
 public class CanNormalAttack : Conditional
 {
+    InGameSceneIfRaidEnd inGameSceneIfRaidEnd;
     InGameSceneCheckTargetAndGetDistance inGameSceneCheckTargetAndGetDistance;
     CharState charState;
-    InGameSceneUiDataManager inGameSceneUiDataManager;
+
 
 
 
 
     public override void OnStart()
     {
-        inGameSceneUiDataManager = GameObject.Find("Manager").gameObject.GetComponent<InGameSceneUiDataManager>();
+        inGameSceneIfRaidEnd = GameObject.Find("Manager").GetComponent<InGameSceneIfRaidEnd>();
         inGameSceneCheckTargetAndGetDistance = gameObject.transform.GetComponent<InGameSceneCheckTargetAndGetDistance>();
         charState = this.gameObject.transform.GetComponent<CharState>();
     }
 
 
     public override TaskStatus OnUpdate()
-    {
-        if (inGameSceneUiDataManager.isBattleStart == false) return TaskStatus.Failure;
+	{
         if (charState.nowState != CharState.NowState.isReadyForAttack ) return TaskStatus.Failure;
-        if (inGameSceneUiDataManager.waitForRaid == true) return TaskStatus.Failure;
+        if (inGameSceneIfRaidEnd.waitForRaid == true) return TaskStatus.Failure;
 
-        if (inGameSceneCheckTargetAndGetDistance.target == null) return TaskStatus.Failure;
+        if (inGameSceneCheckTargetAndGetDistance.target == null)
+        {
+            charState.nowState = CharState.NowState.isIdle;
+            return TaskStatus.Failure;
+        }
 
 
         // 공격 범위 이내일 경우
         if ((inGameSceneCheckTargetAndGetDistance.isEnemyOutOfAttackRange() == false)   
-            && (charState.skillPoint <= 100)) return TaskStatus.Success;
-
-        return TaskStatus.Failure;
+            && (charState.skillPoint <= 100))
+        {
+           // Debug.Log("공격 범위 안에 있다다! + 일반 공격 한다!");
+           return TaskStatus.Success;
+        }
+		return TaskStatus.Failure;
 	}
 }
