+++ Assets/스크립트/BT/가가가가가/CanSpecialAttack.cs
@@ -4,29 +4,33 @@ using BehaviorDesigner.Runtime.Tasks;
 
 public class CanSpecialAttack : Conditional
 {
+    InGameSceneIfRaidEnd inGameSceneIfRaidEnd;
     InGameSceneCheckTargetAndGetDistance inGameSceneCheckTargetAndGetDistance;
     CharState charState;
-    InGameSceneUiDataManager inGameSceneUiDataManager;
+
 
     public override void OnStart()
     {
-        inGameSceneUiDataManager = GameObject.Find("Manager").gameObject.GetComponent<InGameSceneUiDataManager>();
         inGameSceneCheckTargetAndGetDistance = gameObject.transform.GetComponent<InGameSceneCheckTargetAndGetDistance>();
+        inGameSceneIfRaidEnd = GameObject.Find("Manager").GetComponent<InGameSceneIfRaidEnd>();
         charState = gameObject.GetComponent<CharState>();
     }
 
     public override TaskStatus OnUpdate()
-    {
-        if (inGameSceneUiDataManager.isBattleStart == false) return TaskStatus.Failure;
-        if (inGameSceneUiDataManager.waitForRaid == true) return TaskStatus.Failure;
-
-        if ( charState.nowState != CharState.NowState.isReadyForAttack) return TaskStatus.Failure; 
-        if (inGameSceneCheckTargetAndGetDistance.target == null) return TaskStatus.Failure;
-
+	{
+        if ( charState.nowState != CharState.NowState.isReadyForAttack) return TaskStatus.Failure;
+        if (inGameSceneIfRaidEnd.waitForRaid == true) return TaskStatus.Failure;
+        if (inGameSceneCheckTargetAndGetDistance.target == null)
+        {
+            charState.nowState = CharState.NowState.isIdle;
+            return TaskStatus.Failure;
+        }
 
         if ((inGameSceneCheckTargetAndGetDistance.isEnemyOutOfAttackRange() == false) &&
-            (charState.skillPoint >= 100)) return TaskStatus.Success;
-
-        return TaskStatus.Failure;
+            (charState.skillPoint >= 100) )
+        {
+            return TaskStatus.Success;
+        }
+		return TaskStatus.Failure;
 	}
 }
