+++ Assets/스크립트/BT/가가가가가/CheckEnemy.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 
 public class CheckEnemy : Conditional
 {
-
+    InGameSceneIfRaidEnd inGameSceneIfRaidEnd;
     InGameSceneCheckTargetAndGetDistance inGameSceneCheckTargetAndGetDistance;
     InGameSceneUiDataManager inGameSceneUiDataManager;
     InGameSceneCharSpineAniCon inGameSceneCharSpineAniCon;
@@ -17,6 +17,7 @@ public class CheckEnemy : Conditional
 
     public override void OnStart()
     {
+        inGameSceneIfRaidEnd = GameObject.Find("Manager").GetComponent<InGameSceneIfRaidEnd>();
         inGameSceneCheckTargetAndGetDistance = gameObject.GetComponent<InGameSceneCheckTargetAndGetDistance>();
         inGameSceneCharSpineAniCon = gameObject.GetComponent<InGameSceneCharSpineAniCon>();
         inGameSceneUiDataManager = GameObject.Find("Manager").GetComponent<InGameSceneUiDataManager>();
@@ -26,7 +27,7 @@ public class CheckEnemy : Conditional
 
     public override TaskStatus OnUpdate()
 	{
-        if (inGameSceneUiDataManager.waitForRaid == true)
+        if (inGameSceneIfRaidEnd.waitForRaid == true)
         {
             charState.nowState = CharState.NowState.isWalkToOrginPos;
             inGameSceneCharSpineAniCon.run();
@@ -39,7 +40,7 @@ public class CheckEnemy : Conditional
             case "playerChar":
                 if (inGameSceneUiDataManager.enemyObjList.Count == 0 )
                 {
-                    //charState.nowState = CharState.NowState.isIdle;
+                    charState.nowState = CharState.NowState.isIdle;
                     inGameSceneCharSpineAniCon.idle();
                     return TaskStatus.Failure;
                 }
@@ -48,15 +49,17 @@ public class CheckEnemy : Conditional
             case "enemyChar":
                 if (inGameSceneUiDataManager.playerObjList.Count == 0)
                 {
-                    //charState.nowState = CharState.NowState.isIdle;
+                    charState.nowState = CharState.NowState.isIdle;
                     inGameSceneCharSpineAniCon.idle();
                     return TaskStatus.Failure;
                 }
                 break;
         }
 
-        if (pastObj != inGameSceneCheckTargetAndGetDistance.target) charState.nowState = CharState.NowState.isIdle;
-
+        if (pastObj != inGameSceneCheckTargetAndGetDistance.target)
+        {
+            charState.nowState = CharState.NowState.isIdle;
+        }
         pastObj = inGameSceneCheckTargetAndGetDistance.target;
 
         return TaskStatus.Success;
