--- Assets/스크립트/BT/가가가가가/MoveToPosBeforeAttack.cs
@@ -1,44 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using BehaviorDesigner.Runtime;
-using BehaviorDesigner.Runtime.Tasks;
-
-
-public class MoveToPosBeforeAttack : Action
-{
-    CharState charState;
-
-    InGameSceneUiDataManager inGameSceneUiDataManager;
-    InGameSceneCharSpineAniCon inGameSceneCharSpineAniCon;
-    InGameSceneCharMove inGameSceneCharMove;
-    InGameSceneGameManager  inGameSceneGameManager;
-
-
-    public override void OnStart()
-    {
-        inGameSceneGameManager = GameObject.Find("Manager").gameObject.GetComponent<InGameSceneGameManager>();
-        inGameSceneUiDataManager = GameObject.Find("Manager").gameObject.GetComponent<InGameSceneUiDataManager>();
-        inGameSceneCharMove = gameObject.GetComponent<InGameSceneCharMove>();
-        inGameSceneCharSpineAniCon = gameObject.GetComponent<InGameSceneCharSpineAniCon>();
-
-        charState = this.gameObject.transform.GetComponent<CharState>();
-    }
-
-
-    public override TaskStatus OnUpdate()
-    {
-        switch (this.gameObject.tag)
-        {
-            case "playerChar":
-                inGameSceneCharSpineAniCon.run();
-                break;
-
-            case "enemyChar":
-                inGameSceneCharSpineAniCon.idle();
-                break;
-        }
-        inGameSceneCharMove.moveToAttackPos(this.gameObject.tag);
-        return TaskStatus.Success;
-    }
-}
