+++ Assets/스크립트/매니저/인게임/InGameSceneUISceneManager.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class InGameSceneUISceneManager : MonoBehaviour
+{
+  
+}
