+++ Assets/스크립트/매니저/인게임/InGameSceneUISceneManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 11be0915a21c18f4b93ea8789cbd363e
+guid: 065a6583a4e93134cae221c933c192c6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
