+++ Assets/스크립트/매니저/인게임/InGameSceneUiDataManager.cs
@@ -6,9 +6,6 @@ using BehaviorDesigner.Runtime;
 
 public class InGameSceneUiDataManager : MonoBehaviour
 {
-    [SerializeField]
-    GameObject startBattleBtn;
-
     public SkeletonAnimation[] enemySkeletonAnimationSet;
     public ExternalBehaviorTree[] enemyBehaviorTreeDataSet;
     public SkeletonAnimation[] playerSkeletonAnimationSet;
@@ -16,20 +13,15 @@ public class InGameSceneUiDataManager : MonoBehaviour
 
     public GameObject enemyObj;
     public GameObject[] enemyCharInGridPos;
-    public GameObject[] enemyCharMovePosBeforeBattle;
     public List<GameObject> enemyObjList = new List<GameObject>();
 
     public GameObject playerObj;
     public GameObject[] playerCharInGridPos;
-    public GameObject[] playerCharMovePosBeforeBattle;
+
     public List<GameObject> playerObjList = new List<GameObject>();
 
     public string[,,] enemyCharInGrid;
-    
+
     //현재 레이트 수 ( 최대 3 )
     public int battleSceneCount = 0 ;
-
-    public int countPlayerCharArrivedToAttackPos = 0;
-    public bool isBattleStart = false;
-    public bool waitForRaid = false;
 }
